+++ R/bigQuery.R
@@ -20,8 +20,8 @@
 #'   here: \url{https://support.google.com/analytics/answer/3416091?hl=en}
 #' 
 #' 
-#' @param projectId The Google project Id where the BigQuery exports sit.
-#' @param datasetId DatasetId of GA export.  This should match the GA View ID.
+#' @param projectId The Google project Id where the BigQuery exports sit
+#' @param datasetId DatasetId of GA export.  This should match the GA View ID
 #' @param start start date
 #' @param end end date
 #' @param metrics metrics to query
@@ -30,13 +30,16 @@
 #' @param filters filter results
 #' @param max_results How many results to fetch
 #' @param query If query is non-NULL then it will use that and ignore above
-#' @param return_query_only Only return the constructed query, don't call BigQuery.
+#' @param return_query_only Only return the constructed query, don't call BigQuery
+#' @param bucket if over 100000 results, specify a Google Cloud bucket to send data to
+#' @param download_file Where to save asynch files.  If NULL saves to current working directory. 
 #' 
 #' @return data.frame of results
 #' 
 #' @seealso \url{https://support.google.com/analytics/answer/4419694?hl=en}
 #'          \url{https://support.google.com/analytics/answer/3437719?hl=en}
 #' 
+#' @import bigQueryR
 #' @export
 google_analytics_bq <- function(projectId,
                                 datasetId,
@@ -49,20 +52,21 @@ google_analytics_bq <- function(projectId,
                                 # segment=NULL,
                                 max_results=100,
                                 query=NULL,
-                                return_query_only=FALSE){
+                                return_query_only=FALSE,
+                                bucket = NULL,
+                                download_file = NULL){
   
   projectId <- as.character(projectId)
-  datasetID <- as.character(datasetId)
+  datasetId <- as.character(datasetId)
   start <- if(!is.null(start)) as.character(as.Date(start))
   end <- if(!is.null(end)) as.character(as.Date(end))
+  max_results <- as.integer(max_results)
 
   
   if (!requireNamespace("bigQueryR", quietly = TRUE)) {
-    stop("bigQueryR needed for this function to work. Please install it via devtools::install_github('MarkEdmondson1234/bigQueryR')",
+    stop("bigQueryR needed for this function to work. Please install it via install.packages('bigQueryR')",
          call. = FALSE)
   }
-  
-  # bq_tables <- bigQueryR::bqr_list_tables(projectId, datasetId)
   
   ## if Sys.Date() == end then construct for ga_sessions_intradata_ too.
   if(is.null(query)){
@@ -115,12 +119,22 @@ google_analytics_bq <- function(projectId,
     }
   }
   
-  if(max_results < 70000){
+  if(max_results < 100000){
 
     out <- bigQueryR::bqr_query(projectId, datasetId, query)
   } else {
     ## do an async query
-    message("Currently only up to 70,000 rows fetched per call.")
+    if (!requireNamespace("googleCloudStorageR", quietly = TRUE)) {
+      stop("googleCloudStorageR needed for large extracts. Please install it via install.packages('googleCloudStorageR')",
+           call. = FALSE)
+    }
+    
+    message("## Over 100,000 rows to fetch, creating asynchronous query via Google Cloud Storage.")
+    google_analytics_bq_asynch(projectId = projectId,
+                               datasetId = datasetId,
+                               query = query,
+                               bucket = bucket,
+                               download_file = download_file)
     return()
   }
   
@@ -140,6 +154,58 @@ google_analytics_bq <- function(projectId,
   out
 }
 
+#' Asynch fetch
+#' @import bigQueryR
+#' @keywords internal
+google_analytics_bq_asynch <- function(projectId,
+                                       datasetId,
+                                       query,
+                                       bucket,
+                                       download_file){
+  time0 <- Sys.time()
+  required_scopes <- c("https://www.googleapis.com/auth/devstorage.full_control", 
+                       "https://www.googleapis.com/auth/cloud-platform")
+  
+  if(!any(getOption("googleAuthR.scopes.selected") %in% required_scopes)){
+    stop("Need re-authentication with googleAuthR::gar_auth() with one of scopes:", 
+         paste(required_scopes, collapse = " "))
+  }
+  
+  if(is.null(bucket)){
+    ## maybe call googleCloudStorgeR to create bucket?
+    stop("Need a Google Cloud bucket to send data to.  Please create one at https://cloud.google.com/storage/")
+  }
+  
+  tableId <- paste0("googleAnalyticsRjob_", 
+                    gsub("-|:| ","",as.character(Sys.time())), 
+                    "_",
+                    idempotency())
+  
+  query_job <- bigQueryR::bqr_query_asynch(projectId = projectId,
+                                           datasetId = datasetId,
+                                           query = query,
+                                           destinationTableId = tableId)
+  query_job <- bigQueryR::bqr_wait_for_job(query_job)
+  
+  message("\nBigQuery query successful and now in BigQuery tableId: ", tableId,
+          "\n - now extracting data to Cloud Storage bucket", bucket)
+  
+  extract_job <- bigQueryR::bqr_extract_data(projectId = projectId,
+                                             datasetId = datasetId,
+                                             tableId = tableId,
+                                             cloudStorageBucket = bucket)
+  extract_job <- bigQueryR::bqr_wait_for_job(extract_job)
+  
+  message("\nBigQuery extract successful to ", bucket,
+          " - now downloading data from Google Cloud Storage")
+  
+  bigQueryR::bqr_download_extract(extract_job,
+                                  filename = download_file)
+  
+  message("All finished, total job time:", format(difftime(Sys.time(), time0), format = "%H:%M:%S"))
+  
+}
+
 lookup_bq_query_m <- c(visits = "SUM(totals.visits) as sessions",
                        sessions = "SUM(totals.visits) as sessions",
                        pageviews = "SUM(totals.pageviews) as pageviews",
@@ -150,7 +216,8 @@ lookup_bq_query_m <- c(visits = "SUM(totals.visits) as sessions",
                        transactionRevenue = "SUM(totals.transactionRevenue)/1000000 as transactionRevenue",
                        transactionsPerSession = "(SUM(totals.transactions) / SUM(totals.visits)) as transactionsPerSession",
                        revenuePerTransaction = "(SUM(totals.transactionRevenue)/1000000) / SUM(totals.transactions) as revenuePerTransaction",
-                       newVisits = "SUM(totals.newVisits) as newVisits",
+                       newSessions = "SUM(totals.newVisits) as newVisits",
+                       percentNewSessions = "(SUM(totals.newVisits) / SUM(totals.visits))*100 AS percentNewSessions",
                        screenviews = "SUM(totals.screenviews) as screenviews",
                        uniqueScreenviews = "SUM(totals.uniqueScreenviews) as uniqueScreenviews",
                        timeOnScreen = "SUM(totals.timeOnScreen) as timeOnScreen",
@@ -197,7 +264,9 @@ lookup_bq_query_d <- c(referralPath = "trafficSource.referralPath as referralPat
                        pagePath = "hits.page.pagePath as pagePath",
                        eventCategory = "hits.eventInfo.eventCategory as eventCategory",
                        eventAction = "hits.eventInfo.eventAction as eventAction",
-                       eventLabel = "hits.eventInfo.eventLabel as eventLabel")
+                       eventLabel = "hits.eventInfo.eventLabel as eventLabel",
+                       ## from http://www.lunametrics.com/blog/2016/06/23/google-analytics-bigquery-export-schema/
+                       landingPagePath = "FIRST(IF(hits.type = 'PAGE', hits.page.pagePath, NULL)) WITHIN RECORD AS landingPagePath")
 
 
 ## this is hit level: add session and product level too.
@@ -284,4 +353,4 @@ customMetricMaker <- function(customMetricIndex=paste0("metric",1:200)){
 #        customDimensions.value,
 #        NULL)) WITHIN RECORD AS customDimension2,
 # FROM [tableID.ga_sessions_20150305]
-# LIMIT 100 
+# LIMIT 100